requirements.txt
aiohttp==3.3.2
astroid==1.6.5
astunparse-noparen==1.5.6
async-timeout==3.0.0
async2rewrite==0.1.7
asyncio==3.4.3
asyncpg==0.16.0
attrs==18.1.0
beautifulsoup4==4.6.0
bs4==0.0.1
cachetools==2.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
ChatterBot==0.8.7
chatterbot-corpus==1.1.2
cleverbot==2.0.0
cleverwrap==0.2.3.6
click==6.7
colorama==0.3.9
colorlog==3.1.4
cx-Freeze==5.1.1
dblpy==0.1.6
discord==0.0.2
discord.py==1.0.0a1497+g8d39ab8
fortnite==0.0.7
fortnite-python==0.2.5
furl==1.2
future==0.16.0
google==2.0.1
google-auth==1.5.0
google-auth-httplib2==0.0.3
googletrans==2.3.0
httplib2==0.11.3
idna==2.5
idna-ssl==1.1.0
imgurpython==1.1.7
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.2.3
lyricsgenius==0.7.0
mathparse==0.1.1
mccabe==0.6.1
multidict==4.3.1
nekos.py==1.0.1
nltk==3.3
oauthlib==2.1.0
opuslib==3.0.1
orderedmultidict==1.0
osuapi==0.0.30
parsedatetime==2.4
pillow==5.2.0
pipenv==2018.6.25
praw==5.4.0
prawcore==0.14.0
psutil==5.4.6
psycopg2==2.7.5
pyasn1==0.4.3
pyasn1-modules==0.2.2
pychalk==2.0.0
pycparser==2.18
pylint==1.9.2
pymongo==3.7.1
PyNaCl==1.0.1
python-dateutil==2.6.1
python-twitter==3.4.2
PyYAML==3.13
ratelimiter==1.2.0.post0
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
six==1.11.0
SQLAlchemy==1.2.10
tinydb==3.9.0.post1
unicode==2.6
Unidecode==1.0.22
update-checker==0.16
urbandictionary==1.1
uritemplate==3.0.0
urllib3==1.21.1
urllib3-mock==0.3.3
virtualenv==16.0.0
virtualenv-clone==0.3.0
websockets==6.0
wikipedia==1.4.0
wrapt==1.10.11
yapf==0.22.0
yarl==1.2.6
youtube-dl==2018.7.21
